move func to msig.ts
