+++ examples/multisig/scripts/multisig_goal_sc.js
@@ -1,7 +1,7 @@
 /**
  * Description:
- *  This script demonstrates how to
-   - load a signed lsig from assets and use that lsig to validate transactions
+ *  This script demonstrates a transaction validated by an asc signed by multisig account with a threshold of 2
+    User saves the signed logic in assets/ which is extracted and used here.
 */
 const { transferAlgo } = require("./common");
 const { transferMicroAlgos, createMsigAddress } = require("algob");
@@ -11,27 +11,22 @@ async function run(runtimeEnv, deployer) {
   const goldOwnerAccount = deployer.accountsByName.get("gold-owner-account");
   const johnAccount = deployer.accountsByName.get("john-account");
   const bobAccount = deployer.accountsByName.get("bob-account");
-
-  /*
-    This example demonstrates a transaction validated by a smart contract signed by a multisig account with threshold of 2
-    User saves the signed logic signature file in assets/ which is extracted and used here.
-  */
   
   //Generate multi signature account hash 
   const addrs =  [goldOwnerAccount.addr, johnAccount.addr, bobAccount.addr]  // you can replace these addresses with your custom addrs for multisig account.
-  const [mparams, kmdMsigAddress] = createMsigAddress(1, 2, addrs);   // passing (version, threshold, address list)
+  const [mparams, multsigaddr] = createMsigAddress(1, 2, addrs);   // passing (version, threshold, address list)
  
   //Extracting multisig from dissembled file generated by goal (goal clerk compile -D <path_to_file>)
   const signedLsig = await deployer.loadMultiSig("sample-asc.msig", []);  
 
   //Funding multisignature account
-  await transferMicroAlgos(deployer, masterAccount, kmdMsigAddress, 10000000, {note: "Funding multisig account"});
+  await transferMicroAlgos(deployer, masterAccount, multsigaddr, 10000000, {note: "Funding multisig account"});
 
   // Transaction PASS - according to sample-asc.teal logic, amount should be <= 100
-  await transferAlgo(deployer, { addr: kmdMsigAddress }, bobAccount.addr, 58, signedLsig);
+  await transferAlgo(deployer, { addr: multsigaddr }, bobAccount.addr, 58, signedLsig);
 
   // Transaction FAIL - according to sample-asc.teal logic, amount should be <= 100
-  await transferAlgo(deployer, { addr: kmdMsigAddress }, bobAccount.addr, 180, signedLsig);
+  await transferAlgo(deployer, { addr: multsigaddr }, bobAccount.addr, 180, signedLsig);
   
 }
