+++ packages/algob/src/internal/deployer.ts
@@ -3,8 +3,8 @@ import * as algosdk from "algosdk";
 
 import { txWriter } from "../internal/tx-log-writer";
 import { AlgoOperator } from "../lib/algo-operator";
-import { readMsigFromFile } from "../lib/files";
-import { decodeMsigObj, getLsig, logicsig } from "../lib/lsig";
+import { getLsig, logicsig } from "../lib/lsig";
+import { decodeMsigObj, readMsigFromFile } from "../lib/msig";
 import { persistCheckpoint } from "../lib/script-checkpoints";
 import type {
   Account,
